feat: move function to common service
