+++ extensions/iceworks-app/package.json
@@ -314,6 +314,7 @@
     "vscode-test": "^1.3.0"
   },
   "dependencies": {
+    "@iceworks/common-service": "^0.1.0",
     "@types/latest-version": "^4.0.1",
     "fs-extra": "^9.0.0",
     "ice-npm-utils": "^1.4.1",
